+++ src/components/SarManageWidget/Claim/index.tsx
@@ -1,18 +1,15 @@
 import { BigNumber } from '@ethersproject/bignumber';
 import { formatEther } from '@ethersproject/units';
-import numeral from 'numeral';
 import React, { useCallback, useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Box } from 'src/components/Box';
 import { Button } from 'src/components/Button';
 import { Text } from 'src/components/Text';
-import { PNG } from 'src/constants/tokens';
-import { useChainId } from 'src/hooks';
 import { Position, useDerivativeSarClaim } from 'src/state/psarstake/hooks';
+import RewardsInfo from '../Compound/RewardsInfo';
 import ConfirmDrawer from '../ConfirmDrawer';
-import Title from '../Title';
 import { Options } from '../types';
-import { Root, ToolTipText } from './styleds';
+import { Root } from './styleds';
 
 interface Props {
   selectedOption: Options;
@@ -23,14 +20,11 @@ interface Props {
 export default function Claim({ selectedOption, selectedPosition, onChange }: Props) {
   const [openDrawer, setOpenDrawer] = useState(false);
 
-  const chainId = useChainId();
-
   const { attempting, hash, claimError, wrappedOnDismiss, onClaim } = useDerivativeSarClaim(selectedPosition);
 
   const { t } = useTranslation();
 
   const apr = selectedPosition?.apr;
-  const png = PNG[chainId];
 
   const handleConfirmDismiss = useCallback(() => {
     setOpenDrawer(false);
@@ -67,27 +61,12 @@ export default function Claim({ selectedOption, selectedPosition, onChange }: Pr
   return (
     <Box>
       <Root>
-        <Title selectedOption={selectedOption} onChange={onChange} />
-        {!selectedPosition ? (
-          <Box>
-            <Text color="text1" fontSize="24px" fontWeight={500} textAlign="center">
-              {t('sarStakeMore.choosePosition')}
-            </Text>
-          </Box>
-        ) : (
-          <Box>
-            <Text color="text1" fontSize="16px" fontWeight={500} textAlign="center">
-              {t('sarCompound.reward')}:
-            </Text>
-            <ToolTipText color="text1" fontSize="36px" fontWeight={500} textAlign="center">
-              {numeral(formatEther(pendingRewards.toString())).format('0.00a')}
-              <span className="tooltip">
-                {formatEther(pendingRewards.toString())} {png.symbol}
-              </span>
-            </ToolTipText>
-          </Box>
-        )}
-
+        <RewardsInfo
+          selectedOption={selectedOption}
+          onChange={onChange}
+          pendingRewards={formatEther(pendingRewards.toString())}
+          selectedPosition={selectedPosition}
+        />
         <Box display="grid" bgColor="color3" borderRadius="4px" padding="20px" style={{ gridGap: '20px' }}>
           <Box display="flex" justifyContent="space-between">
             <Box>
